Make SUBSCRIBEs stateless
To make the agent BLF works when switching from A to B,
we configured the phones to send their SUBSCRIBES at a
very short interval. This allows the REGISTER and SUBSCRIBE
to "follow" each other and get the SUBSCRIBE to the
appropriate stack as soon as possible.

If we keep the SUBSCRIBEs within the dialog the new
SUBSCRIBEs are still being send to the innactive stack,
preventing the active one from getting the subscription
and sending its NOTIFYies